I'm trying to figure out or remember under what circumstances the old code to find the first changeOnNewBranch was needed.
If p4 interterchange returns nothing, it means that the new branch was created at the head of the base branch and we do not need to roll back any commits in the main git branch.  If there are changes in the base branch that we don't want in the new child branch, p4 interchange would return true.

OK, I think the use-case was:

* child branch is created (at the head of the base branch, at that point in time) and a change was made on it, so now the branch could be created in the git repo
* before the new branch was created in git, another change (with a higher changelist number) was created on the main branch
* now the import was done on the child branch, and we need to reject the change from the base branch with the higher changelist number

But this scenario should be captured by the p4 interchange command, I would think.  So I believe that old code should be completely removed.  It's not as though this code path of the script is well-tested or well-used.